Dev ana

+++ magiadoslivros/src/main/java/com/eternos/magiadoslivros/domain/repository/FornecedorRepository.java
@@ -4,6 +4,6 @@ import com.eternos.magiadoslivros.domain.model.Fornecedor;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FornecedorRepository extends JpaRepository<Fornecedor, Long>{
+public interface FornecedorRepository extends JpaRepository<Fornecedor, Integer>{
     
 }
